Удалил файл с моками
