+++ src/const.js
@@ -5,3 +5,56 @@ export const sortTypes = {
   TIME: `Time`,
   PRICE: `Price`
 };
+
+export const pointTypes = {
+  'taxi': {
+    name: `taxi`,
+    group: `transfer`,
+    emoji: `0x1F695`
+  },
+  'bus': {
+    name: `bus`,
+    group: `transfer`,
+    emoji: `0x1F68C`
+  },
+  'train': {
+    name: `train`,
+    group: `transfer`,
+    emoji: `0x1F682`
+  },
+  'ship': {
+    name: `ship`,
+    group: `transfer`,
+    emoji: `0x1F6F3`
+  },
+  'transport': {
+    name: `transport`,
+    group: `transfer`,
+    emoji: `0x1F68A`
+  },
+  'drive': {
+    name: `drive`,
+    group: `transfer`,
+    emoji: `0x1F697`
+  },
+  'flight': {
+    name: `flight`,
+    group: `transfer`,
+    emoji: `0x2708`
+  },
+  'check-in': {
+    name: `check-in`,
+    group: `activity`,
+    emoji: `0x1F3E8`
+  },
+  'sightseeing': {
+    name: `sightseeing`,
+    group: `activity`,
+    emoji: `0x1F3DB`
+  },
+  'restaurant': {
+    name: `restaurant`,
+    group: `activity`,
+    emoji: `0x1F374`
+  }
+};
